Resolve dependency cycles between `actions-util` and `workflow`
