+++ lib/actions-util.test.js
@@ -30,14 +30,15 @@ const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const ava_1 = __importDefault(require("ava"));
 const sinon = __importStar(require("sinon"));
-const actionsutil = __importStar(require("./actions-util"));
+const actionsUtil = __importStar(require("./actions-util"));
+const api_client_1 = require("./api-client");
 const environment_1 = require("./environment");
 const testing_utils_1 = require("./testing-utils");
 const util_1 = require("./util");
 (0, testing_utils_1.setupTests)(ava_1.default);
 (0, ava_1.default)("getRef() throws on the empty string", async (t) => {
     process.env["GITHUB_REF"] = "";
-    await t.throwsAsync(actionsutil.getRef);
+    await t.throwsAsync(actionsUtil.getRef);
 });
 (0, ava_1.default)("getRef() returns merge PR ref if GITHUB_SHA still checked out", async (t) => {
     await (0, util_1.withTmpDir)(async (tmpDir) => {
@@ -46,9 +47,9 @@ const util_1 = require("./util");
         const currentSha = "a".repeat(40);
         process.env["GITHUB_REF"] = expectedRef;
         process.env["GITHUB_SHA"] = currentSha;
-        const callback = sinon.stub(actionsutil, "getCommitOid");
+        const callback = sinon.stub(actionsUtil, "getCommitOid");
         callback.withArgs("HEAD").resolves(currentSha);
-        const actualRef = await actionsutil.getRef();
+        const actualRef = await actionsUtil.getRef();
         t.deepEqual(actualRef, expectedRef);
         callback.restore();
     });
@@ -60,10 +61,10 @@ const util_1 = require("./util");
         process.env["GITHUB_REF"] = expectedRef;
         process.env["GITHUB_SHA"] = "b".repeat(40);
         const sha = "a".repeat(40);
-        const callback = sinon.stub(actionsutil, "getCommitOid");
+        const callback = sinon.stub(actionsUtil, "getCommitOid");
         callback.withArgs("refs/remotes/pull/1/merge").resolves(sha);
         callback.withArgs("HEAD").resolves(sha);
-        const actualRef = await actionsutil.getRef();
+        const actualRef = await actionsUtil.getRef();
         t.deepEqual(actualRef, expectedRef);
         callback.restore();
     });
@@ -73,10 +74,10 @@ const util_1 = require("./util");
         (0, testing_utils_1.setupActionsVars)(tmpDir, tmpDir);
         process.env["GITHUB_REF"] = "refs/pull/1/merge";
         process.env["GITHUB_SHA"] = "a".repeat(40);
-        const callback = sinon.stub(actionsutil, "getCommitOid");
+        const callback = sinon.stub(actionsUtil, "getCommitOid");
         callback.withArgs(tmpDir, "refs/pull/1/merge").resolves("a".repeat(40));
         callback.withArgs(tmpDir, "HEAD").resolves("b".repeat(40));
-        const actualRef = await actionsutil.getRef();
+        const actualRef = await actionsUtil.getRef();
         t.deepEqual(actualRef, "refs/pull/1/head");
         callback.restore();
     });
@@ -84,16 +85,16 @@ const util_1 = require("./util");
 (0, ava_1.default)("getRef() returns ref provided as an input and ignores current HEAD", async (t) => {
     await (0, util_1.withTmpDir)(async (tmpDir) => {
         (0, testing_utils_1.setupActionsVars)(tmpDir, tmpDir);
-        const getAdditionalInputStub = sinon.stub(actionsutil, "getOptionalInput");
+        const getAdditionalInputStub = sinon.stub(actionsUtil, "getOptionalInput");
         getAdditionalInputStub.withArgs("ref").resolves("refs/pull/2/merge");
         getAdditionalInputStub.withArgs("sha").resolves("b".repeat(40));
         // These values are be ignored
         process.env["GITHUB_REF"] = "refs/pull/1/merge";
         process.env["GITHUB_SHA"] = "a".repeat(40);
-        const callback = sinon.stub(actionsutil, "getCommitOid");
+        const callback = sinon.stub(actionsUtil, "getCommitOid");
         callback.withArgs("refs/pull/1/merge").resolves("b".repeat(40));
         callback.withArgs("HEAD").resolves("b".repeat(40));
-        const actualRef = await actionsutil.getRef();
+        const actualRef = await actionsUtil.getRef();
         t.deepEqual(actualRef, "refs/pull/2/merge");
         callback.restore();
         getAdditionalInputStub.restore();
@@ -107,7 +108,7 @@ const util_1 = require("./util");
         process.env["CODE_SCANNING_REF"] = expectedRef;
         process.env["GITHUB_REF"] = "";
         process.env["GITHUB_SHA"] = currentSha;
-        const actualRef = await actionsutil.getRef();
+        const actualRef = await actionsUtil.getRef();
         t.deepEqual(actualRef, expectedRef);
     });
 });
@@ -119,17 +120,17 @@ const util_1 = require("./util");
         process.env["CODE_SCANNING_REF"] = "refs/pull/1/HEAD";
         process.env["GITHUB_REF"] = expectedRef;
         process.env["GITHUB_SHA"] = currentSha;
-        const actualRef = await actionsutil.getRef();
+        const actualRef = await actionsUtil.getRef();
         t.deepEqual(actualRef, expectedRef);
     });
 });
 (0, ava_1.default)("getRef() throws an error if only `ref` is provided as an input", async (t) => {
     await (0, util_1.withTmpDir)(async (tmpDir) => {
         (0, testing_utils_1.setupActionsVars)(tmpDir, tmpDir);
-        const getAdditionalInputStub = sinon.stub(actionsutil, "getOptionalInput");
+        const getAdditionalInputStub = sinon.stub(actionsUtil, "getOptionalInput");
         getAdditionalInputStub.withArgs("ref").resolves("refs/pull/1/merge");
         await t.throwsAsync(async () => {
-            await actionsutil.getRef();
+            await actionsUtil.getRef();
         }, {
             instanceOf: Error,
             message: "Both 'ref' and 'sha' are required if one of them is provided.",
@@ -141,10 +142,10 @@ const util_1 = require("./util");
     await (0, util_1.withTmpDir)(async (tmpDir) => {
         (0, testing_utils_1.setupActionsVars)(tmpDir, tmpDir);
         process.env["GITHUB_WORKSPACE"] = "/tmp";
-        const getAdditionalInputStub = sinon.stub(actionsutil, "getOptionalInput");
+        const getAdditionalInputStub = sinon.stub(actionsUtil, "getOptionalInput");
         getAdditionalInputStub.withArgs("sha").resolves("a".repeat(40));
         await t.throwsAsync(async () => {
-            await actionsutil.getRef();
+            await actionsUtil.getRef();
         }, {
             instanceOf: Error,
             message: "Both 'ref' and 'sha' are required if one of them is provided.",
@@ -153,19 +154,19 @@ const util_1 = require("./util");
     });
 });
 (0, ava_1.default)("computeAutomationID()", async (t) => {
-    let actualAutomationID = actionsutil.computeAutomationID(".github/workflows/codeql-analysis.yml:analyze", '{"language": "javascript", "os": "linux"}');
+    let actualAutomationID = (0, api_client_1.computeAutomationID)(".github/workflows/codeql-analysis.yml:analyze", '{"language": "javascript", "os": "linux"}');
     t.deepEqual(actualAutomationID, ".github/workflows/codeql-analysis.yml:analyze/language:javascript/os:linux/");
     // check the environment sorting
-    actualAutomationID = actionsutil.computeAutomationID(".github/workflows/codeql-analysis.yml:analyze", '{"os": "linux", "language": "javascript"}');
+    actualAutomationID = (0, api_client_1.computeAutomationID)(".github/workflows/codeql-analysis.yml:analyze", '{"os": "linux", "language": "javascript"}');
     t.deepEqual(actualAutomationID, ".github/workflows/codeql-analysis.yml:analyze/language:javascript/os:linux/");
     // check that an empty environment produces the right results
-    actualAutomationID = actionsutil.computeAutomationID(".github/workflows/codeql-analysis.yml:analyze", "{}");
+    actualAutomationID = (0, api_client_1.computeAutomationID)(".github/workflows/codeql-analysis.yml:analyze", "{}");
     t.deepEqual(actualAutomationID, ".github/workflows/codeql-analysis.yml:analyze/");
     // check non string environment values
-    actualAutomationID = actionsutil.computeAutomationID(".github/workflows/codeql-analysis.yml:analyze", '{"number": 1, "object": {"language": "javascript"}}');
+    actualAutomationID = (0, api_client_1.computeAutomationID)(".github/workflows/codeql-analysis.yml:analyze", '{"number": 1, "object": {"language": "javascript"}}');
     t.deepEqual(actualAutomationID, ".github/workflows/codeql-analysis.yml:analyze/number:/object:/");
     // check undefined environment
-    actualAutomationID = actionsutil.computeAutomationID(".github/workflows/codeql-analysis.yml:analyze", undefined);
+    actualAutomationID = (0, api_client_1.computeAutomationID)(".github/workflows/codeql-analysis.yml:analyze", undefined);
     t.deepEqual(actualAutomationID, ".github/workflows/codeql-analysis.yml:analyze/");
 });
 (0, ava_1.default)("initializeEnvironment", (t) => {
@@ -175,7 +176,7 @@ const util_1 = require("./util");
 (0, ava_1.default)("isAnalyzingDefaultBranch()", async (t) => {
     process.env["GITHUB_EVENT_NAME"] = "push";
     process.env["CODE_SCANNING_IS_ANALYZING_DEFAULT_BRANCH"] = "true";
-    t.deepEqual(await actionsutil.isAnalyzingDefaultBranch(), true);
+    t.deepEqual(await actionsUtil.isAnalyzingDefaultBranch(), true);
     process.env["CODE_SCANNING_IS_ANALYZING_DEFAULT_BRANCH"] = "false";
     await (0, util_1.withTmpDir)(async (tmpDir) => {
         (0, testing_utils_1.setupActionsVars)(tmpDir, tmpDir);
@@ -188,18 +189,18 @@ const util_1 = require("./util");
         process.env["GITHUB_EVENT_PATH"] = envFile;
         process.env["GITHUB_REF"] = "main";
         process.env["GITHUB_SHA"] = "1234";
-        t.deepEqual(await actionsutil.isAnalyzingDefaultBranch(), true);
+        t.deepEqual(await actionsUtil.isAnalyzingDefaultBranch(), true);
         process.env["GITHUB_REF"] = "refs/heads/main";
-        t.deepEqual(await actionsutil.isAnalyzingDefaultBranch(), true);
+        t.deepEqual(await actionsUtil.isAnalyzingDefaultBranch(), true);
         process.env["GITHUB_REF"] = "feature";
-        t.deepEqual(await actionsutil.isAnalyzingDefaultBranch(), false);
+        t.deepEqual(await actionsUtil.isAnalyzingDefaultBranch(), false);
         fs.writeFileSync(envFile, JSON.stringify({
             schedule: "0 0 * * *",
         }));
         process.env["GITHUB_EVENT_NAME"] = "schedule";
         process.env["GITHUB_REF"] = "refs/heads/main";
-        t.deepEqual(await actionsutil.isAnalyzingDefaultBranch(), true);
-        const getAdditionalInputStub = sinon.stub(actionsutil, "getOptionalInput");
+        t.deepEqual(await actionsUtil.isAnalyzingDefaultBranch(), true);
+        const getAdditionalInputStub = sinon.stub(actionsUtil, "getOptionalInput");
         getAdditionalInputStub
             .withArgs("ref")
             .resolves("refs/heads/something-else");
@@ -208,7 +209,7 @@ const util_1 = require("./util");
             .resolves("0000000000000000000000000000000000000000");
         process.env["GITHUB_EVENT_NAME"] = "schedule";
         process.env["GITHUB_REF"] = "refs/heads/main";
-        t.deepEqual(await actionsutil.isAnalyzingDefaultBranch(), false);
+        t.deepEqual(await actionsUtil.isAnalyzingDefaultBranch(), false);
         getAdditionalInputStub.restore();
     });
 });
@@ -222,9 +223,9 @@ const util_1 = require("./util");
         process.env["GITHUB_REPOSITORY"] = "octocat/HelloWorld";
         process.env["CODEQL_ACTION_ANALYSIS_KEY"] = "analysis-key";
         process.env["RUNNER_OS"] = "macOS";
-        const getRequiredInput = sinon.stub(actionsutil, "getRequiredInput");
+        const getRequiredInput = sinon.stub(actionsUtil, "getRequiredInput");
         getRequiredInput.withArgs("matrix").resolves("input/matrix");
-        const statusReport = await actionsutil.createStatusReportBase("init", "failure", new Date("May 19, 2023 05:19:00"), "failure cause", "exception stack trace");
+        const statusReport = await (0, api_client_1.createStatusReportBase)("init", "failure", new Date("May 19, 2023 05:19:00"), "failure cause", "exception stack trace");
         t.assert(typeof statusReport.job_run_uuid === "string");
         t.assert(statusReport.workflow_run_id === 100);
         t.assert(statusReport.workflow_run_attempt === 2);
