+++ lib/analyze-action-env.test.js
@@ -47,7 +47,7 @@ const util = __importStar(require("./util"));
         process.env["GITHUB_REPOSITORY"] = "github/codeql-action-fake-repository";
         process.env["GITHUB_API_URL"] = "https://api.github.com";
         sinon
-            .stub(actionsUtil, "createStatusReportBase")
+            .stub(api, "createStatusReportBase")
             .resolves({});
         sinon.stub(api, "sendStatusReport").resolves(true);
         sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
