+++ lib/analyze-action.js
@@ -50,7 +50,7 @@ const util = __importStar(require("./util"));
 const util_1 = require("./util");
 async function sendStatusReport(startedAt, config, stats, error, trapCacheUploadTime, dbCreationTimings, didUploadTrapCaches, logger) {
     const status = actionsUtil.getActionsStatus(error, stats?.analyze_failure_language);
-    const statusReportBase = await actionsUtil.createStatusReportBase("finish", status, startedAt, error?.message, error?.stack);
+    const statusReportBase = await api.createStatusReportBase("finish", status, startedAt, error?.message, error?.stack);
     const statusReport = {
         ...statusReportBase,
         ...(config
@@ -146,7 +146,7 @@ async function run() {
     util.initializeEnvironment(actionsUtil.getActionVersion());
     const logger = (0, logging_1.getActionsLogger)();
     try {
-        if (!(await api.sendStatusReport(await actionsUtil.createStatusReportBase("finish", "starting", startedAt)))) {
+        if (!(await api.sendStatusReport(await api.createStatusReportBase("finish", "starting", startedAt)))) {
             return;
         }
         config = await (0, config_utils_1.getConfig)(actionsUtil.getTemporaryDirectory(), logger);
