+++ lib/resolve-environment-action.js
@@ -39,7 +39,7 @@ async function run() {
     const logger = (0, logging_1.getActionsLogger)();
     const language = (0, languages_1.resolveAlias)((0, actions_util_1.getRequiredInput)("language"));
     try {
-        if (!(await (0, api_client_1.sendStatusReport)(await (0, actions_util_1.createStatusReportBase)(ACTION_NAME, "starting", startedAt)))) {
+        if (!(await (0, api_client_1.sendStatusReport)(await (0, api_client_1.createStatusReportBase)(ACTION_NAME, "starting", startedAt)))) {
             return;
         }
         const gitHubVersion = await (0, api_client_1.getGitHubVersion)();
@@ -63,11 +63,11 @@ async function run() {
         else {
             // For any other error types, something has more seriously gone wrong and we fail.
             core.setFailed(`Failed to resolve a build environment suitable for automatically building your code. ${error.message}`);
-            await (0, api_client_1.sendStatusReport)(await (0, actions_util_1.createStatusReportBase)(ACTION_NAME, (0, actions_util_1.getActionsStatus)(error), startedAt, error.message, error.stack));
+            await (0, api_client_1.sendStatusReport)(await (0, api_client_1.createStatusReportBase)(ACTION_NAME, (0, actions_util_1.getActionsStatus)(error), startedAt, error.message, error.stack));
         }
         return;
     }
-    await (0, api_client_1.sendStatusReport)(await (0, actions_util_1.createStatusReportBase)(ACTION_NAME, "success", startedAt));
+    await (0, api_client_1.sendStatusReport)(await (0, api_client_1.createStatusReportBase)(ACTION_NAME, "success", startedAt));
 }
 async function runWrapper() {
     try {
