+++ lib/workflow.js
@@ -26,7 +26,7 @@ var __importDefault = (this && this.__importDefault) || function (mod) {
     return (mod && mod.__esModule) ? mod : { "default": mod };
 };
 Object.defineProperty(exports, "__esModule", { value: true });
-exports.getCheckoutPathInputOrThrow = exports.getUploadInputOrThrow = exports.getCategoryInputOrThrow = exports.getWorkflowRunAttempt = exports.getWorkflowRunID = exports.getWorkflowRelativePath = exports.getWorkflow = exports.formatWorkflowCause = exports.formatWorkflowErrors = exports.validateWorkflow = exports.getWorkflowErrors = exports.WorkflowErrors = exports.patternIsSuperset = void 0;
+exports.getCheckoutPathInputOrThrow = exports.getUploadInputOrThrow = exports.getCategoryInputOrThrow = exports.getWorkflow = exports.formatWorkflowCause = exports.formatWorkflowErrors = exports.validateWorkflow = exports.getWorkflowErrors = exports.WorkflowErrors = exports.patternIsSuperset = void 0;
 const fs = __importStar(require("fs"));
 const path = __importStar(require("path"));
 const zlib_1 = __importDefault(require("zlib"));
@@ -185,7 +185,7 @@ exports.getWorkflow = getWorkflow;
  * Get the absolute path of the currently executing workflow.
  */
 async function getWorkflowAbsolutePath(logger) {
-    const relativePath = await getWorkflowRelativePath();
+    const relativePath = await api.getWorkflowRelativePath();
     const absolutePath = path.join((0, util_1.getRequiredEnvParam)("GITHUB_WORKSPACE"), relativePath);
     if (fs.existsSync(absolutePath)) {
         logger.debug(`Derived the following absolute path for the currently executing workflow: ${absolutePath}.`);
@@ -195,55 +195,6 @@ async function getWorkflowAbsolutePath(logger) {
         "This can happen if the currently running workflow checks out a branch that doesn't contain " +
         "the corresponding workflow file.");
 }
-/**
- * Get the path of the currently executing workflow relative to the repository root.
- */
-async function getWorkflowRelativePath() {
-    const repo_nwo = (0, util_1.getRequiredEnvParam)("GITHUB_REPOSITORY").split("/");
-    const owner = repo_nwo[0];
-    const repo = repo_nwo[1];
-    const run_id = Number((0, util_1.getRequiredEnvParam)("GITHUB_RUN_ID"));
-    const apiClient = api.getApiClient();
-    const runsResponse = await apiClient.request("GET /repos/:owner/:repo/actions/runs/:run_id?exclude_pull_requests=true", {
-        owner,
-        repo,
-        run_id,
-    });
-    const workflowUrl = runsResponse.data.workflow_url;
-    const workflowResponse = await apiClient.request(`GET ${workflowUrl}`);
-    return workflowResponse.data.path;
-}
-exports.getWorkflowRelativePath = getWorkflowRelativePath;
-/**
- * Get the workflow run ID.
- */
-function getWorkflowRunID() {
-    const workflowRunIdString = (0, util_1.getRequiredEnvParam)("GITHUB_RUN_ID");
-    const workflowRunID = parseInt(workflowRunIdString, 10);
-    if (Number.isNaN(workflowRunID)) {
-        throw new Error(`GITHUB_RUN_ID must define a non NaN workflow run ID. Current value is ${workflowRunIdString}`);
-    }
-    if (workflowRunID < 0) {
-        throw new Error(`GITHUB_RUN_ID must be a non-negative integer. Current value is ${workflowRunIdString}`);
-    }
-    return workflowRunID;
-}
-exports.getWorkflowRunID = getWorkflowRunID;
-/**
- * Get the workflow run attempt number.
- */
-function getWorkflowRunAttempt() {
-    const workflowRunAttemptString = (0, util_1.getRequiredEnvParam)("GITHUB_RUN_ATTEMPT");
-    const workflowRunAttempt = parseInt(workflowRunAttemptString, 10);
-    if (Number.isNaN(workflowRunAttempt)) {
-        throw new Error(`GITHUB_RUN_ATTEMPT must define a non NaN workflow run attempt. Current value is ${workflowRunAttemptString}`);
-    }
-    if (workflowRunAttempt <= 0) {
-        throw new Error(`GITHUB_RUN_ATTEMPT must be a positive integer. Current value is ${workflowRunAttemptString}`);
-    }
-    return workflowRunAttempt;
-}
-exports.getWorkflowRunAttempt = getWorkflowRunAttempt;
 function getStepsCallingAction(job, actionName) {
     if (job.uses) {
         throw new Error(`Could not get steps calling ${actionName} since the job calls a reusable workflow.`);
