+++ src/analyze-action-env.test.ts
@@ -27,7 +27,7 @@ test("analyze action with RAM & threads from environment variables", async (t) =
     process.env["GITHUB_REPOSITORY"] = "github/codeql-action-fake-repository";
     process.env["GITHUB_API_URL"] = "https://api.github.com";
     sinon
-      .stub(actionsUtil, "createStatusReportBase")
+      .stub(api, "createStatusReportBase")
       .resolves({} as actionsUtil.StatusReportBase);
     sinon.stub(api, "sendStatusReport").resolves(true);
     sinon.stub(actionsUtil, "isAnalyzingDefaultBranch").resolves(true);
